requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
python-dotenv>=1.0
sqlalchemy>=2.0
pymysql>=1.1
joblib>=1.3
numpy>=1.26
pandas>=2.0
scikit-learn==1.4.2
lightgbm>=4.0
python-dateutil>=2.8
